+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
-All notable changes to this project will be documented in this file.## 0.0.1 (05.06.2020)
+All notable changes to this project will be documented in this file.
+## 0.2.0 (010.06.2020)
+- stored max. Power Production too the file
+- config.host noe use 'envoy.local' path or Your configured iP adress
+
 ## 0.1.1 (08.06.2020)
 - many fixes and more info in log